refactor: scanfamilies config creation
