+++ runtime_scan/pkg/config/config.go
@@ -25,6 +25,7 @@ import (
 	"github.com/spf13/viper"
 
 	"github.com/openclarity/vmclarity/runtime_scan/pkg/config/aws"
+	"github.com/openclarity/vmclarity/shared/pkg/families"
 )
 
 const (
@@ -35,19 +36,10 @@ const (
 	DeleteJobPolicy                 = "DELETE_JOB_POLICY"
 	ScannerContainerImage           = "SCANNER_CONTAINER_IMAGE"
 	ScannerKeyPairName              = "SCANNER_KEY_PAIR_NAME"
-	GitleaksBinaryPath              = "GITLEAKS_BINARY_PATH"
-	ClamBinaryPath                  = "CLAM_BINARY_PATH"
-	FreshclamBinaryPath             = "FRESHCLAM_BINARY_PATH"
-	AlternativeFreshclamMirrorURL   = "ALTERNATIVE_FRESHCLAM_MIRROR_URL"
-	LynisInstallPath                = "LYNIS_INSTALL_PATH"
 	AttachedVolumeDeviceName        = "ATTACHED_VOLUME_DEVICE_NAME"
 	defaultAttachedVolumeDeviceName = "xvdh"
 	ScannerBackendAddress           = "SCANNER_VMCLARITY_BACKEND_ADDRESS"
 	ScanConfigWatchInterval         = "SCAN_CONFIG_WATCH_INTERVAL"
-	ExploitDBAddress                = "EXPLOIT_DB_ADDRESS"
-	TrivyServerAddress              = "TRIVY_SERVER_ADDRESS"
-	GrypeServerAddress              = "GRYPE_SERVER_ADDRESS"
-	ChkrootkitBinaryPath            = "CHKROOTKIT_BINARY_PATH"
 )
 
 type OrchestratorConfig struct {
@@ -70,11 +62,8 @@ type ScannerConfig struct {
 	// address to use.
 	ScannerBackendAddress string
 
-	ExploitsDBAddress string
-
-	TrivyServerAddress string
-
-	GrypeServerAddress string
+	// Scanner server and exploit DB addresses for families
+	FamiliesAddresses families.Addresses
 
 	JobResultTimeout          time.Duration
 	JobResultsPollingInterval time.Duration
@@ -90,23 +79,8 @@ type ScannerConfig struct {
 	// Mainly used for debugging.
 	ScannerKeyPairName string
 
-	// The gitleaks binary path in the scanner image container.
-	GitleaksBinaryPath string
-
-	// The clam binary path in the scanner image container.
-	ClamBinaryPath string
-
-	// The freshclam binary path in the scanner image container
-	FreshclamBinaryPath string
-
-	// The freshclam mirror url to use if it's enabled
-	AlternativeFreshclamMirrorURL string
-
-	// The location where Lynis is installed in the scanner image
-	LynisInstallPath string
-
-	// The chkrootkit binary path in the scanner image container.
-	ChkrootkitBinaryPath string
+	// The families paths
+	FamiliesPaths families.Paths
 
 	// the name of the block device to attach to the scanner job
 	DeviceName string
@@ -119,16 +93,7 @@ func setConfigDefaults(backendHost string, backendPort int, backendBaseURL strin
 	viper.SetDefault(ScanConfigWatchInterval, "30s")
 	viper.SetDefault(DeleteJobPolicy, string(DeleteJobPolicyAlways))
 	viper.SetDefault(ScannerBackendAddress, fmt.Sprintf("http://%s%s", net.JoinHostPort(backendHost, strconv.Itoa(backendPort)), backendBaseURL))
-	// https://github.com/openclarity/vmclarity-tools-base/blob/main/Dockerfile#L33
-	viper.SetDefault(GitleaksBinaryPath, "/artifacts/gitleaks")
-	// https://github.com/openclarity/vmclarity-tools-base/blob/main/Dockerfile#L35
-	viper.SetDefault(LynisInstallPath, "/artifacts/lynis")
-	// https://github.com/openclarity/vmclarity-tools-base/blob/main/Dockerfile
-	viper.SetDefault(ChkrootkitBinaryPath, "/artifacts/chkrootkit")
-	viper.SetDefault(ExploitDBAddress, fmt.Sprintf("http://%s", net.JoinHostPort(backendHost, "1326")))
 	viper.SetDefault(AttachedVolumeDeviceName, defaultAttachedVolumeDeviceName)
-	viper.SetDefault(ClamBinaryPath, "clamscan")
-	viper.SetDefault(FreshclamBinaryPath, "freshclam")
 
 	viper.AutomaticEnv()
 }
@@ -140,24 +105,17 @@ func LoadConfig(backendHost string, backendPort int, baseURL string) (*Orchestra
 		AWSConfig:             aws.LoadConfig(),
 		ScannerBackendAddress: viper.GetString(ScannerBackendAddress),
 		ScannerConfig: ScannerConfig{
-			Region:                        viper.GetString(ScannerAWSRegion),
-			JobResultTimeout:              viper.GetDuration(JobResultTimeout),
-			JobResultsPollingInterval:     viper.GetDuration(JobResultsPollingInterval),
-			ScanConfigWatchInterval:       viper.GetDuration(ScanConfigWatchInterval),
-			DeleteJobPolicy:               getDeleteJobPolicyType(viper.GetString(DeleteJobPolicy)),
-			ScannerImage:                  viper.GetString(ScannerContainerImage),
-			ScannerBackendAddress:         viper.GetString(ScannerBackendAddress),
-			ScannerKeyPairName:            viper.GetString(ScannerKeyPairName),
-			GitleaksBinaryPath:            viper.GetString(GitleaksBinaryPath),
-			LynisInstallPath:              viper.GetString(LynisInstallPath),
-			DeviceName:                    viper.GetString(AttachedVolumeDeviceName),
-			ExploitsDBAddress:             viper.GetString(ExploitDBAddress),
-			ClamBinaryPath:                viper.GetString(ClamBinaryPath),
-			FreshclamBinaryPath:           viper.GetString(FreshclamBinaryPath),
-			AlternativeFreshclamMirrorURL: viper.GetString(AlternativeFreshclamMirrorURL),
-			TrivyServerAddress:            viper.GetString(TrivyServerAddress),
-			GrypeServerAddress:            viper.GetString(GrypeServerAddress),
-			ChkrootkitBinaryPath:          viper.GetString(ChkrootkitBinaryPath),
+			Region:                    viper.GetString(ScannerAWSRegion),
+			JobResultTimeout:          viper.GetDuration(JobResultTimeout),
+			JobResultsPollingInterval: viper.GetDuration(JobResultsPollingInterval),
+			ScanConfigWatchInterval:   viper.GetDuration(ScanConfigWatchInterval),
+			DeleteJobPolicy:           getDeleteJobPolicyType(viper.GetString(DeleteJobPolicy)),
+			ScannerImage:              viper.GetString(ScannerContainerImage),
+			ScannerBackendAddress:     viper.GetString(ScannerBackendAddress),
+			ScannerKeyPairName:        viper.GetString(ScannerKeyPairName),
+			DeviceName:                viper.GetString(AttachedVolumeDeviceName),
+			FamiliesPaths:             families.LoadPaths(),
+			FamiliesAddresses:         families.LoadAddresses(backendHost),
 		},
 	}
